requirements.txt
anthropic
python-dotenv
sounddevice
numpy
SpeechRecognition
pyttsx3
pyaudio
